requirements.txt
Flask==2.3.3
qrcode==7.4.2
pillow==10.0.1
firebase-admin==6.2.0
